+++ awint_macros/src/lib.rs
@@ -6,11 +6,12 @@
 //! ## Scope Dependencies
 //!
 //! All of the macros often require `Bits` and `InlAwi` to be in scope. This
-//! could be from `awint_core` or be a reexport from `awint`. `extawi!` and
-//! similar always requires `ExtAwi` to be in scope, which could be imported
-//! from `awint_ext` or be a reexport from `awint`. `cc!` may require none,
-//! one, or both storage types depending on the input. Macros with fallible
-//! inputs require `Option<T>` variants to be in scope.
+//! could be from `awint_core` or be a reexport from `awint`. `extawi!` always
+//! requires `ExtAwi` to be in scope, `awi!` always requires `Awi` to be in
+//! scope, etc., which could be imported from `awint_ext` or be a reexport from
+//! `awint`. `cc!` may require any of these storage types depending on the
+//! input. Macros with fallible inputs require `Option<T>` variants to be in
+//! scope.
 //!
 //! ## Concatenations of Components
 //!
@@ -227,8 +228,9 @@
 //! Anything that has well defined `bw() -> usize`, `Deref::deref() -> &Bits`,
 //! `DerefMut::deref_mut() -> &mut Bits`, `AsRef::as_ref() -> &Bits`, and
 //! `AsMut::as_mut() -> &mut Bits` functions can be used as a variable.
-//! Arbitrary `Bits` references, `ExtAwi`, `InlAwi`, `FP<B>` and other well
-//! defined arbitrary width integer types can thus be used as variables.
+//! Arbitrary `Bits` references, the `*Awi` types, wrapper types like `FP<B>`,
+//! and other well defined arbitrary width integer types can thus be used as
+//! variables.
 //!
 //! ```
 //! use awint::awi::*;
@@ -236,20 +238,23 @@
 //! let source = inlawi!(0xc4di64);
 //! // a bunch of zeroed 64 bit arbitrary width integers from different
 //! // storage types and construction methods.
-//! let mut awi: ExtAwi = ExtAwi::zero(bw(64));
-//! let mut a: &mut Bits = awi.const_as_mut();
+//! let mut tmp: ExtAwi = ExtAwi::zero(bw(64));
+//! let mut a: &mut Bits = tmp.const_as_mut();
 //! let mut b: ExtAwi = extawi!(0i64);
 //! let mut c: inlawi_ty!(64) = <inlawi_ty!(64)>::zero();
-//! let mut d: FP<inlawi_ty!(64)> = FP::new(false, inlawi!(0u64), 32).unwrap();
+//! let mut d: Awi = Awi::zero_with_capacity(bw(64), bw(128));
+//! let mut e: FP<inlawi_ty!(64)> = FP::new(false, inlawi!(0u64), 32).unwrap();
 //!
-//! // Use the `cc` macro to copy the source concatenation `source` to the sink
-//! // concatenations `a`, `b`, `c`, and `d`. Here, every concatenation is just
+//! // Use the `cc` macro to copy the source concatenation to the
+//! // sink concatenations. Here, every concatenation is just
 //! // a single variable component.
-//! cc!(source; a; b; c; d).unwrap();
+//! cc!(source; a; b; c; d; e).unwrap();
 //!
 //! assert_eq!(a, inlawi!(0xc4di64).as_ref());
 //! assert_eq!(b, extawi!(0xc4di64));
 //! assert_eq!(c, inlawi!(0xc4di64));
+//! assert_eq!(d, awi!(0xc4di64));
+//! assert_eq!(e, FP::new(false, inlawi!(0xc4di64), 32).unwrap());
 //!
 //! let awi = inlawi!(0xau4);
 //! let a = awi.as_ref();
@@ -270,23 +275,23 @@
 //! ```
 //! use awint::awi::*;
 //!
-//! let y3 = inlawi!(0xba9u12);
-//! let y2 = inlawi!(0x876u12);
-//! let y1 = inlawi!(0x543u12);
-//! let y0 = inlawi!(0x210u12);
+//! let y3 = awi!(0xba9u12);
+//! let y2 = awi!(0x876u12);
+//! let y1 = awi!(0x543u12);
+//! let y0 = awi!(0x210u12);
 //!
-//! let mut z2 = inlawi!(0u16);
-//! let mut z1 = inlawi!(0u16);
-//! let mut z0 = inlawi!(0u16);
+//! let mut z2 = awi!(0u16);
+//! let mut z1 = awi!(0u16);
+//! let mut z0 = awi!(0u16);
 //!
 //! cc!(
 //!     y3, y2, y1, y0;
 //!     z2, z1, z0;
 //! ).unwrap();
 //!
-//! assert_eq!(z2, inlawi!(0xba98u16));
-//! assert_eq!(z1, inlawi!(0x7654u16));
-//! assert_eq!(z0, inlawi!(0x3210u16));
+//! assert_eq!(z2, awi!(0xba98u16));
+//! assert_eq!(z1, awi!(0x7654u16));
+//! assert_eq!(z0, awi!(0x3210u16));
 //! ```
 //!
 //! Visually, the components in the two concatenations are being aligned like
@@ -311,12 +316,12 @@
 //! // with 0 (e.x. `0..r`) can have the zero omitted and just use `..r`.
 //! // Ranges ending like `r..` will include all bits up to the most
 //! // significant bit of the variable.
-//! let b = extawi!(
+//! let b = awi!(
 //!     a[..=7], a[(a.bw() - 16)..];
 //!     a[(5 * 4)..(9 * 4)], a[..(2 * 4)];
 //! ).unwrap();
 //! assert_eq!(a, inlawi!(0x9109843276u40));
-//! assert_eq!(b, extawi!(0x109876_u24));
+//! assert_eq!(b, awi!(0x109876_u24));
 //! ```
 //!
 //! Also note: if you see the compiler complain about mutability of `Bits`
@@ -388,35 +393,35 @@
 //!
 //! // filler bits in source concatenations have no effects and sink bits are
 //! // preserved
-//! let x = inlawi!(0xabcd_u16);
+//! let x = awi!(0xabcd_u16);
 //!
-//! let mut y = inlawi!(0x123456_u24);
+//! let mut y = awi!(0x123456_u24);
 //! cc!(
 //!     x, ..8;
 //!     y;
 //! ).unwrap();
-//! assert_eq!(y, inlawi!(0xabcd56_u24));
+//! assert_eq!(y, awi!(0xabcd56_u24));
 //!
-//! y = inlawi!(0x123456_u24);
+//! y = awi!(0x123456_u24);
 //! cc!(
 //!     ..8, x;
 //!     y;
 //! ).unwrap();
-//! assert_eq!(y, inlawi!(0x12abcd_u24));
+//! assert_eq!(y, awi!(0x12abcd_u24));
 //!
 //! // filler bits in sink concatenations can act as spacers that drop the
 //! // effects of source bits, note that the effect is processed independently
 //! // for every sink concatenation.
-//! let x = inlawi!(0x123456_u24);
-//! let mut y = inlawi!(0u16);
-//! let mut z = inlawi!(0u16);
+//! let x = awi!(0x123456_u24);
+//! let mut y = awi!(0u16);
+//! let mut z = awi!(0u16);
 //! cc!(
 //!     x;
 //!     y, ..8;
 //!     ..8, z;
 //! ).unwrap();
-//! assert_eq!(y, inlawi!(0x1234_u16));
-//! assert_eq!(z, inlawi!(0x3456_u16));
+//! assert_eq!(y, awi!(0x1234_u16));
+//! assert_eq!(z, awi!(0x3456_u16));
 //! ```
 //!
 //! Fillers are also useful in cases where all concatenations lack a needed
@@ -435,7 +440,7 @@
 //! // but the macro will be able to determine that the common width is 44.
 //! assert_eq!(inlawi!(x; ..44).unwrap(), inlawi!(-99i44));
 //!
-//! // `extawi!` and `cc!` can also run into the problem with unbounded
+//! // the other macros can also run into the problem with unbounded
 //! // fillers, which will be explained later. We can additionally use a
 //! // dynamic range (but be sure to always use a static filler concatenation
 //! // if possible so that internal buffers can be on the stack).
@@ -468,10 +473,10 @@
 //!
 //! // we introduce the initialization specifier, which can be one of "zero",
 //! // "umax", "imax", "imin", or "uone" corresponding to the standard
-//! // construction functions of `ExtAwi` and `InlAwi` (and you can add your
-//! // own initializations by implementing traits for `InlAwi` and `ExtAwi`
-//! // with functions matching the formats of the `InlAwi` constructors and the
-//! // hidden `panicking_` functions on `ExtAwi`)
+//! // construction functions of a `*Awi` type (and you can add your
+//! // own initializations by implementing traits for the `*Awi` type
+//! // with functions matching the formats of the `*Awi` constructors and the
+//! // hidden `panicking_` functions on the dynamic with `ExtAwi` and `Awi`)
 //! let x = extawi!(umax: ..8);
 //! assert_eq!(x, ExtAwi::umax(bw(8)));
 //!
@@ -661,17 +666,17 @@
 //!   unexpected behavior. The parser and code generator treats identical
 //!   strings like they produce the same value every time, and would call `f()`
 //!   only once.
-//! - In general, the macros use the `Bits::field` operation to copy different
-//!   bitfields independently to a buffer, then field from the buffer to the
-//!   sink components. When concatenations take the form `variable or constant
-//!   with full range; var_1[..]; var_2[..]; var_3[..], ...`, the macros use
-//!   `Bits::copy_` to directly copy without an intermediate buffer. This copy
-//!   assigning mode cannot copy between `Bits` references that point to the
-//!   same underlying storage, because it results in aliasing. Thus, trying to
-//!   do something like `cc!(x; x)` results in the borrow checker complaining
-//!   about macro generated variables within the macro being borrowed as both
-//!   immutable and mutable. `cc!(x; x)` is semantically a no-op anyway, so it
-//!   should not be used.
+//! - In the most general case, the macros use the `Bits::field` operation to
+//!   copy different bitfields independently to a buffer, then field from the
+//!   buffer to the sink components. When concatenations take the form
+//!   `variable or constant with full range; var_1[..]; var_2[..]; var_3[..],
+//!   ...`, the macros use `Bits::copy_` to directly copy without an
+//!   intermediate buffer. This copy assigning mode cannot copy between `Bits`
+//!   references that point to the same underlying storage, because it results
+//!   in aliasing. Thus, trying to do something like `cc!(x; x)` results in the
+//!   borrow checker complaining about macro generated variables within the
+//!   macro being borrowed as both immutable and mutable. `cc!(x; x)` is
+//!   semantically a no-op anyway, so it should not be used.
 //! - The code generated by the macros takes special care to avoid panicking or
 //!   overflowing. If a range is reversed (the start value is larger than the
 //!   end value), the macro will will return a compile time error or `None` at
@@ -689,7 +694,7 @@
 extern crate proc_macro;
 use awint_macro_internals::{
     awint_macro_bits, awint_macro_cc, awint_macro_extawi, awint_macro_inlawi,
-    unstable_native_inlawi_ty,
+    unstable_native_inlawi_ty, awint_macro_awi,
 };
 use proc_macro::TokenStream;
 
@@ -742,6 +747,17 @@ pub fn inlawi(input: TokenStream) -> TokenStream {
 #[proc_macro]
 pub fn extawi(input: TokenStream) -> TokenStream {
     match awint_macro_extawi(&input.to_string()) {
+        Ok(s) => s.parse().unwrap(),
+        Err(s) => panic!("{}", s),
+    }
+}
+
+/// A concatenations of components macro, additionally using the source value to
+/// construct an [Awi](awint_macro_internals::awint_ext::Awi). See the
+/// [crate documentation](crate) for more.
+#[proc_macro]
+pub fn awi(input: TokenStream) -> TokenStream {
+    match awint_macro_awi(&input.to_string()) {
         Ok(s) => s.parse().unwrap(),
         Err(s) => panic!("{}", s),
     }